adds swim-lane-placement sketch, but this is not how we should do it. Instead we should be utilizing a single flex implementation for the masonry layout, and set a breaking element.

+++ scripts/dev/generate-leaderboards.js
@@ -44,7 +44,9 @@ function fixUpName(inputString) {
 }
 
 const tables = await dbAll(db, 'PRAGMA show_tables;');
-fs.rmSync('./src/routes/dev/leaderboard/data/',  { recursive: true });
+if (fs.existsSync('./src/routes/dev/leaderboard/data/')) {
+    fs.rmSync('./src/routes/dev/leaderboard/data/',  { recursive: true });
+}
 fs.mkdirSync('./src/routes/dev/leaderboard/data/');
 for (let { name } of tables) {
     // get column names.
